Map DWORD to uint instead of int, according to https://msdn.microsoft.com/en-us/library/cc230318.aspx

+++ ManagedCertificates/MarshalExtensions.cs
@@ -6,10 +6,10 @@ namespace ManagedCertificates
     public static class MarshalExtensions
     {
         // This is needed because structs with variable-sized arrays cannot be marshaled directly
-        public static string[] PtrToStringArray(IntPtr pointer, int arrayLength)
+        public static string[] PtrToStringArray(IntPtr pointer, uint arrayLength)
         {
             IntPtr[] outPointers = new IntPtr[arrayLength];
-            Marshal.Copy(pointer, outPointers, 0, arrayLength);
+            Marshal.Copy(pointer, outPointers, 0, (int)arrayLength);
 
             var result = new string[arrayLength];
             for (int i = 0; i < arrayLength; i++)
